LU-12477 lustre: use delete_from_page_cache() for directory pages.
lustre sometimes uses the internal function truncate_complete_page()
to remove a page of a directory.
Much of what this function does, does not apply to directory pages
as there is no invalidatepage function, and at these times, the
page is not dirty.
The only useful part of the function is delete_from_page_cache(),
so just call that directly.

Linux-commit: d17fa2f3a0b9b40be48e0c3cc88eb3b3cea1b701

Change-Id: I54795e71e107c50f662bd2015c6f621bfe436e0a
Acked-by: Oleg Drokin <green@linuxhacker.ru>
Signed-off-by: NeilBrown <neilb@suse.com>
Reviewed-on: https://review.whamcloud.com/35462
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Lai Siyao <lai.siyao@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>