LU-12137 osd-ldiskfs: shared common code for osd lookup
For osd_lookup_one_len_unlocked() the only time we don't return
the found dentry is if is_bad_inode() is true. The test for
is_bad_inode() can only be run when d_inode is not NULL so
we can move the d_inode test. The other change is the drop of
the debug message which will just fill the logs with too much
chatter so remove it.

Change-Id: I82cf73ca842a45d906ffc21c9c5397a61c2679d8
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/35581
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Hongchao Zhang <hongchao@whamcloud.com>
Reviewed-by: Neil Brown <neilb@suse.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>