LU-9897 utils: remove dependency on libblkid
There was a weak dependency in llverdev on libblkid in order to
find the size of the underlying block device.  However, in the
absence of libblkid there is fallback code that can find the
block device size, so we don't really need libblkid at all.

Similarly, llverdev had a weak dependency on libext2fs to check
if the block device was currently in use, but all kernels since
v2.6.34-rc4-22-g6b4517a7913a allow using open(O_EXCL) to ensure
a block device is not mounted, so there is no need for llverdev
to depend on libext2fs for this.

Remove the conditional code depending on the external libraries
from llverdev and the corresponding configuration checks, which
were causing problems in various build environments.

Test-Parameters: trivial testlist=conf-sanity
Fixes: 6bc323747706 ("LU-9897 build: use pkgconf for detecting libblkid")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I90d2f0c99115ce07168c185dec6628be813ebbe5
Reviewed-on: https://review.whamcloud.com/36010
Reviewed-by: Patrick Farrell <pfarrell@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>